+++ src/app/common/components/Icon/styles.ts
