+++ src/app/common/components/Modal/styles.ts
