+++ src/app/common/components/Notification/styles.ts
