+++ src/app/common/components/Toolbar/styles.ts
