+++ src/app/common/types.ts
