+++ src/app/modules/landing/components/PostCard/styles.ts
