+++ src/app/modules/landing/components/PostsList/styles.ts
