+++ src/app/modules/viewPost/components/Comment/index.tsx
@@ -1,0 +1,14 @@
+import { IPostComment } from 'src/app/modules/viewPost/types';
+
+interface ICommentProps {
+	comment: IPostComment;
+}
+
+export default function Comment({ comment }: ICommentProps) {
+	return (
+		<div key={comment.id}>
+			<p>{comment.comment}</p>
+			<small>{comment.userId}</small>
+		</div>
+	);
+}
