+++ src/app/modules/viewPost/types.ts
