+++ src/pages/_app.tsx
@@ -1,7 +1,7 @@
 import { AppProps } from 'next/app';
 import React from 'react';
-import Notification from 'src/common/components/Notification';
-import AppTheme from 'src/common/themes/AppTheme';
+import Notification from 'src/app/common/components/Notification';
+import AppTheme from 'src/app/common/themes/AppTheme';
 
 export default function App({ Component, pageProps }: AppProps) {
 	return (
