+++ src/pages/index.ts
@@ -1,3 +1,3 @@
-import LandingPage from 'src/features/landing';
+import LandingPage from 'src/app/modules/landing';
 
 export default LandingPage;
